Rename krate::metadata::krate to show_minimal
