+++ src/tests/krate/mod.rs
@@ -1,7 +1,6 @@
 mod dependencies;
 mod downloads;
 mod following;
-mod krate;
 mod owners;
 mod publish;
 mod reverse_dependencies;
